+++ Changelog.md
@@ -3,3 +3,8 @@
 Added Title Bar
 ### Version 0.1.2
 Removed Decoration or frame around the window
+### ~~Version 0.1.2.1~~ 
+~~Minor Patch just setting default height to 600 and 800~~
+### Version 0.1.3
+Major Overhaul visualy of the title bar
+**Issue**: Tauri Data Drag region still doesnt work as of this update